Rename errors.js to alert.js.
