+++ src/core/App.vue
@@ -34,7 +34,7 @@ import NotReady from './shell/NotReady.vue';
 import { refreshCache } from './idb/cache';
 import { loadFieldModules, updateFieldModules } from './field-modules';
 import bundles, { loadConfigDocs, updateConfigDocs } from './store/configDocuments';
-import { alert } from './store/errors';
+import { alert } from './store/alert';
 import profile, { loadProfile, updateProfile } from './store/profile';
 import settings from './store/settings';
 import flattenEntity from './utils/flattenEntity';
