+++ src/core/index.js
@@ -7,7 +7,7 @@ import { setRouter } from './store/useRouter';
 import App from './App.vue';
 import AppBarOptions from './shell/AppBarOptions.vue';
 import mountFieldModule from './field-modules/mount';
-import { alert } from './store/errors';
+import { alert } from './store/alert';
 import t from './mixins/t';
 import farm from './farm';
 import utils from './utils';
